feat: Refactor YamlSyntaxValidationService for improved exception handling and add YamlDotNetDiagnosticFactory for diagnostic creation
